refactor: move `training/` and change imports
to `functionals/training/`
