+++ plantseg/functionals/prediction/utils/utils.py
@@ -4,7 +4,7 @@ from plantseg import PATH_PREDICT_TEMPLATE
 from plantseg.core.zoo import model_zoo
 from plantseg.functionals.prediction.utils.array_dataset import ArrayDataset
 from plantseg.functionals.prediction.utils.slice_builder import SliceBuilder
-from plantseg.training.augs import get_test_augmentations
+from plantseg.functionals.training.augs import get_test_augmentations
 from plantseg.utils import load_config
 
 logger = logging.getLogger(__name__)
