+++ plantseg/functionals/training/embeddings.py
